Add ad logic and premium upgrade service with clean architecture
