+++ lib/providers/ad_provider.dart
@@ -1,0 +1,26 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:appoint/providers/user_subscription_provider.dart';
+
+// Provider that determines if ads should be shown
+final shouldShowAdsProvider = Provider<bool>((ref) {
+  final subscriptionAsync = ref.watch(userSubscriptionProvider);
+  
+  // Check subscription status - if user is premium or has admin access, don't show ads
+  return subscriptionAsync.maybeWhen(
+    data: (isPremium) => !isPremium,
+    orElse: () => true, // Show ads by default if subscription status is unknown
+  );
+});
+
+// Provider for premium upgrade functionality (placeholder)
+final premiumUpgradeProvider = Provider<PremiumUpgradeService>((ref) {
+  return PremiumUpgradeService();
+});
+
+class PremiumUpgradeService {
+  Future<void> upgradeToPremium() async {
+    // TODO: Implement real payment flow
+    // This would integrate with Stripe, in-app purchases, etc.
+    // For now, just a placeholder
+  }
+}
