+++ lib/widgets/booking_confirmation_sheet.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:appoint/providers/user_subscription_provider.dart';
+import 'package:appoint/providers/ad_provider.dart';
 import 'package:appoint/services/ad_service.dart';
 
 class BookingConfirmationSheet extends ConsumerStatefulWidget {
